BodyID is an int

+++ EliteJournalReader/Events/ScanEvent.cs
@@ -194,7 +194,7 @@ namespace EliteJournalReader.Events
 
             public string BodyName { get; set; }
 
-            public long BodyID { get; set; }
+            public int BodyID { get; set; }
 
             public double DistanceFromArrivalLs { get; set; }
 
@@ -283,7 +283,7 @@ namespace EliteJournalReader.Events
     public struct BodyParent
     {
         public string Type;
-        public long BodyID;
+        public int BodyID;
     }
 
 
@@ -310,7 +310,7 @@ namespace EliteJournalReader.Events
                             var bp = new BodyParent
                             {
                                 Type = prop.Name,
-                                BodyID = prop.Value.Value<long>()
+                                BodyID = prop.Value.Value<int>()
                             };
                             bps.Add(bp);
                         }
